# HorusVis/frontend/shared/HorusVisWeb/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HorusVis API",
    "description": "Scaffold API for HorusVis.",
    "version": "v1"
  },
  "paths": {
...
    "/api/reports/dashboard": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDashboardDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDashboardDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDashboardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/bug-density": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BugDensityItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BugDensityItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BugDensityItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/team-performance": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamPerformanceItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamPerformanceItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamPerformanceItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/critical-issues": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "topN",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CriticalIssueDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CriticalIssueDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CriticalIssueDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/recommendations": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/export": {
      "post": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "csv"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BugDensityItemDto": {
        "type": "object",
        "properties": {
          "featureArea": {
            "type": "string",
            "nullable": true
          },
          "openCount": {
            "type": "integer",
            "format": "int32"
          },
          "resolvedCount": {
            "type": "integer",
            "format": "int32"
          },
          "avgTimeToCloseHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CriticalIssueDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "issueCode": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "nullable": true
          },
          "severity": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "assigneeName": {
            "type": "string",
            "nullable": true
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RecommendationItemDto": {
        "type": "object",
        "properties": {
          "ruleKey": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReportDashboardDto": {
        "type": "object",
        "properties": {
          "totalActiveBugs": {
            "type": "integer",
            "format": "int32"
          },
          "avgTimeToCloseHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "taskVelocityPoints": {
            "type": "number",
            "format": "double"
          },
          "criticalPriorityCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalActiveBugsTrend": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "avgTimeToCloseTrend": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "taskVelocityTrend": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "criticalPriorityTrend": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TeamPerformanceItemDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "tasksCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "totalPoints": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Paste a bearer token returned from POST /api/auth/login.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": [ ]
    }
  ]
}